==> 02_Scripts/09_fitting_GAM_and_statistics.Rmd <==
---
title: "GAM fitting of archive & monitoring data "
subtitle: "MUC GoF & Gotland & Landsort"
author: "Juliane Romahn"
date: "07.12.23"
output:
  pdf_document: default
  html_document: default
  word_document: default
editor_options:
  chunk_output_type: inline
papersize: a4
---

```{r}
 rm(list = ls())
library(tidyverse)
library(ggplot2)
library(ggpubr)
#library(readxl)
#library(viridis)
library(mgcv)
#library(nord)
library(ggExtra)
library(caret)
# install.packages("devtools")
#devtools::install_github("G-Thomson/Manu")
library(Manu)

```


```{r setup}
# has to be in setup & is only correct outside of the setup
knitr::opts_knit$set(root.dir = '/PATH/TO/02_Scripts_folder')
getwd()
```

```{r}
workspace <- "04_workspace_2025"
output_path <- "04_results_2025"
output_path_extra <- file.path(output_path, "GAM")
if (!dir.exists(output_path_extra)){ dir.create(output_path_extra) }
source("02_Scripts/00_R_functions.R") # read file with functions
```


```{r}
hoiho2 <-c("#44781E","#1F59C7", "#9ABBD5", "#746455")
load(file.path(workspace,"Data_allApproach__for_modelling.RData"))

total_data.yearly <- total_data.yearly_norm %>%
    mutate(method=gsub("Dinosterl.dinostanols..µg.gTOC.","Dinosterl+dinostanols (µg/gTOC)", method))%>%
    mutate(method=gsub("C25.1.C27.1..µg.gTOC.","C25:1+C27:1 (µg/gTOC)", method))


rm(total_data.yearly_norm)
```

# first identify good k

## stats -- gam
```{r}
k_values <- c(15,20,25,30, 35,40,45, 50)
methods <- c("REML", "ML")
basis <- c("tp", "cr", "gp")
set.seed(12345)
N <- 300 # number of points at which to evaluate the smooth
list_k_plots <- list()
k_stats <- data.frame()
colors_list <- c("#44781E", "#2C3B75") ##


for(my_method in methods){
  for (my_bas in basis){
    for ( meth in unique(total_data.yearly$method)){
      print (meth)
      subset_1 <- total_data.yearly %>% filter( method == meth)
      plot_list_method <- list()
      for (org in unique( subset_1$organism) ){
        subset_2 <- subset_1 %>% filter ( organism == org)
        for (loc in unique(subset_2$location2 )){
          subset_3 <- subset_2 %>%  filter( location2 == loc)%>%
                          mutate( age= as.numeric(age),
                                  values=as.numeric(values)) %>%
                      na.omit()
    
          for (my_k in k_values){
            if(length(subset_3$values) > my_k){
              #define color
              color_plot <- ""
              if(org == "Phytoplankton"){color_plot = colors_list[1]}
              else{color_plot=colors_list[2]}
              
              m <- gam( values ~ s(age, k =my_k, bs= my_bas), data=subset_3 , method = my_method)
    
              sum_stats <- summary(m)
              k.stats <- as.data.frame(k.check(m)) # k-index close to 1 nice, and edf is NOT close to k'
              k.stats <- k.stats %>% mutate(location = loc, 
                                        method= meth, 
                                        organism = org, 
                                        df= (length(subset_3$values)-1), 
                                        gam_method= my_method,
                                        basis= my_bas,
                                        PaCoeff_es=sum_stats$p.coeff,
                                        PaCoeff_t= sum_stats$p.t,
                                        PaCoeff_p.value=sum_stats$p.pv,
                                        PaCoeff_p=round(sum_stats$p.pv,5),
                                        sm_f.value=sum_stats$s.table[3],
                                        sm_p.value=sum_stats$s.table[4],
                                        sm_p=round(sum_stats$s.table[4],5),
                                        dev.expl=sum_stats$dev.expl,
                                        weights= "no weights")
                
              k_stats <- rbind(k_stats, k.stats)
              
            }
          }
        }
      }
    }
  }
}

colnames(k_stats)
k_stats <- k_stats %>% arrange( method, location,"k'", organism, gam_method)

k_stats %>% group_by(method, location, organism) %>%
  summarise(maxK= max(`k'`)+1)


```
## select model parameters
```{r}
select <- k_stats %>% filter(`k-index`> 0.8 & `p-value` >0.05)%>%
  group_by(location, method, organism)%>%
  filter(`k-index` ==max(`k-index`))#%>%
  #filter(`k-index` >=max(round(`k-index`,1)))
write.table(select, 
            file = file.path(output_path_extra, "Parameters_GAM_modeling.tsv"), sep = "\t", row.names = F)


selected <- select %>%
    select(location, method, organism,`k'`, gam_method, basis)%>%
    mutate(k_value=`k'`+1 )
```

## second fit model and predict data with determined k

```{r}
#total_data.yearly
my_predicted <- data.frame()

##### 
for ( loc in unique(total_data.yearly$location2 )){
  subset_1 <- total_data.yearly %>% filter(location2 == loc )
  min_monitoring <- min( subset_1 %>% filter(method=="Biomonitoring")%>% pull(age))
   # extract smallest highest age for the different methods
  max_monitoring <- min( subset_1 %>% group_by(method)%>% summarize(max_value = max(age))%>% pull(max_value))
  
   # extract smallest highest age for the different methods
  min_MUC <- min( subset_1 %>% filter(method!="Biomonitoring")%>% group_by(method) %>% summarise(age = min(age))%>% pull(age))
  max_MUC <- min( subset_1 %>% filter(method!="Biomonitoring")%>% group_by(method) %>% summarise(age =max(age))%>% pull(age))
  timeframe_MUC <-with(subset_3, data.frame(age = seq(min(min_MUC), max(max_MUC),
                                              length.out = (as.numeric(max(max_MUC))- as.numeric(min(min_MUC)))))) # 30
  
  timeframe_moni <- with(subset_1, data.frame(age = seq(min(min_monitoring), max(max_monitoring),
                                              length.out = as.numeric(max(max_monitoring))- as.numeric(min(min_monitoring))))) # 10

  for (meth in unique(subset_1$method) ){
    subset_2 <- subset_1 %>% filter ( method == meth)
    for (org in unique( subset_2$organism)){
      subset_3 <- subset_2 %>%  filter( organism == org)%>%
                      mutate( age= as.numeric(age),
                              values=as.numeric(values)) %>%
                  na.omit()
      
      
      ## get model parameters
      selected_row <- selected %>% filter( method== meth & location == loc & organism == org)
      my_method <- selected_row$gam_method
      my_bas <- selected_row$basis 
      my_k <- selected_row$k_value
      
      ####### fit and predict model
      m <- gam( values ~ s(age, k =my_k, bs= my_bas), data=subset_3 , method = my_method)
      
      ## add GAM GCV results. (monitoring timeframe)
      newBraya <- timeframe_moni %>% filter(age >= min(subset_3$age))
      fit_gcv <- predict(m, newdata = newBraya, se.fit = TRUE) 
      crit.t <- qt(0.975, df.residual(m))
      newGCV <- data.frame(age = newBraya[["age"]],
                               fit= fit_gcv$fit,
                                se.fit = fit_gcv$se.fit)
      newBraya <- transform(newGCV, 
                              upper = fit + (crit.t * se.fit),
                              lower =fit - (crit.t * se.fit))
          
          
          ## Add indicator variable for model
      newBraya <- transform(newBraya, method = rep( "GAM (GCV)", length(newBraya$age)),
                                      dataset = rep(meth, length(newBraya$age)),
                                      k= rep(my_k, length(newBraya$age)),
                                      gam_meth = rep(my_method, length(newBraya$age)),
                                      gam_basis = rep(my_bas, length(newBraya$age)),
                                      organism= rep(org, length(newBraya$age)),
                                      location= rep(loc, length(newBraya$age)),
                                      timespan= rep("Monitoring", length(newBraya$age)),
                                      timespan.det= rep(paste("Monitoring:",min_monitoring, "-", max_monitoring), length(newBraya$age)))
      my_predicted <- rbind(my_predicted,newBraya)
      
      if(meth!="Biomonitoring"){
        ## add GAM GCV results. (MUC timeframe)

        newBraya <-   timeframe_MUC %>% filter(age >= min(subset_3$age))
        fit_gcv <- predict(m, newdata = newBraya, se.fit = TRUE) 
        crit.t <- qt(0.975, df.residual(m))
        newGCV <- data.frame(age = newBraya[["age"]],
                                 fit= fit_gcv$fit,
                                  se.fit = fit_gcv$se.fit)
        newBraya <- transform(newGCV, 
                                upper = fit + (crit.t * se.fit),
                                lower =fit - (crit.t * se.fit))
            
            
        ## Add indicator variable for model
        newBraya <- transform(newBraya, method = rep( "GAM (GCV)", length(newBraya$age)),
                                        dataset = rep(meth, length(newBraya$age)),
                                        k= rep(my_k, length(newBraya$age)),
                                        gam_meth = rep(my_method, length(newBraya$age)),
                                        gam_basis = rep(my_bas, length(newBraya$age)),
                                        organism= rep(org, length(newBraya$age)),
                                        location= rep(loc, length(newBraya$age)),
                                        timespan= rep("Archive", length(newBraya$age)),
                                        timespan.det= rep(paste("Archive:",min_MUC, "-", max_MUC), length(newBraya$age)))
        my_predicted <- rbind(my_predicted,newBraya)
      }
    }
  }
}
```


## calulate spearman correlation for monitoring timeframe and for MUC archive time frame for GAM DATA!!!
### monitoring time span compare against monitoring, MUC compare against metabarcoding

```{r}
predict_for_stats <- my_predicted%>% 
  mutate(se.fit=NULL, upper= NULL, lower=NULL, k = NULL, gam_meth=NULL, gam_basis=NULL, timespan.det) %>%
  pivot_wider(names_from = dataset, values_from = fit)
stats_gam_predicted <- data.frame()

scatter_plots <- c()

stats_meth = "spearman"
for(org in rev(unique(predict_for_stats$organism))){
  print(org)
  subset_1 <- predict_for_stats %>% filter(organism ==org)
  for( time in unique(subset_1$timespan)){
    subset_2 <- subset_1 %>% filter(timespan ==time) #time
    for(loc in unique(subset_2$location)){
      subset_3 <- subset_2 %>% filter(location ==loc) #time
      
      # set up comparison
      loop_through <- c()
      if(grepl("Monitoring", time)){
        loop_through <- c("Metabarcoding")
        main_approach="Biomonitoring"
        if(loc !="LD"){
          if (org =="Dinoflagellata" ){loop_through <- c(loop_through,"Dinosterl+dinostanols (µg/gTOC)")
          }else  {loop_through <- c(loop_through,c("C25:1+C27:1 (µg/gTOC)"))}
        }
      }else if(grepl("Archive", time) & loc !="LD"){ ### for longtime
          main_approach="Metabarcoding"
          loop_through <- c()
          if (org =="Dinoflagellata"){loop_through <- c(loop_through,"Dinosterl+dinostanols (µg/gTOC)")
          }else { loop_through <- c(loop_through ,c("C25:1+C27:1 (µg/gTOC)"))}
      }
      #### calculcating
      for(method1 in loop_through ){
        search_term <- gsub("\\+", "\\\\+", method1)
        search_term <- gsub("\\(.*\\)", "", search_term)
        search_term <- gsub("^", "", search_term)
        index <- grep(search_term, colnames(subset_3))
        index2 <- grep(main_approach, colnames(subset_3))
        
        data <- data.frame(x=as.numeric(unlist(subset_3[,index] )), y=as.numeric(unlist(subset_3[,index2] ))  )
        data <- data %>%na.omit()
        
        # o stats and add to table
        stats = cor.test(data$x,data$y , method = stats_meth)
        stats_gam_predicted <- rbind(stats_gam_predicted, data.frame(
          organism= org, location=unique(subset_3$location), timespan=time, 
          df= (length(subset_3$location)-1),
          Method1=method1, Method2= main_approach, dataset= unique(subset_3$method),
          stats_method=stats_meth, estimate=stats$estimate, statistic =stats$statistic, p.value=stats$p.value))
        
        plot_stats <- ggplot(data, aes(x=x, y=y)) +
          geom_point() +
          geom_smooth(method=lm , color="red", fill="#69b3a2", se=TRUE)+
          stat_cor(method=stats_meth, geom= "label")+
          labs(x=method1, y= main_approach, title= paste( org, unique(subset_2$location), sep=" - ") )
        scatter_plots <- c( scatter_plots, list(plot_stats))
        
        if (grepl("MUC", time) ){
          ## repeat for data after 1950
          subset_4 <- subset_3 %>% filter(age > 1950)
          data <- data.frame(x=as.numeric(unlist(subset_4[,index] )), y=subset_4$Metabarcoding  )
          
          stats = cor.test(data$x, data$y, method = stats_meth)
          stats_gam_predicted <- rbind(stats_gam_predicted, data.frame(
            organism= org, location=unique(subset_3$location), timespan="after 1950", 
            df= (length(subset_4$Metabarcoding)-1),
            Method1=method1, Method2= main_approach, dataset= unique(subset_4$method), 
            stats_method=stats_meth, estimate=stats$estimate, statistic =stats$statistic, p.value=stats$p.value))
          
          plot_stats <- ggplot(data, aes(x=x, y=y)) +
            geom_point() +
            geom_smooth(method=lm , color="red", fill="#69b3a2", se=TRUE)+
            stat_cor(method=stats_meth, geom= "label")+
            labs(x=method1, y= main_approach, title= paste( org, "-" , loc, "(>1950)", sep=" ") )
          scatter_plots <- c( scatter_plots, list(plot_stats))
        }
      }
    }
  }
}
stats_plot <- ggarrange(plotlist=scatter_plots, ncol = 4, nrow=5)
plot(stats_plot)
ggsave(stats_plot, 
      file = file.path(output_path_extra, "Correlation_results_plots.png"), width=14, height=14)
ggsave(stats_plot, 
      file = file.path(output_path_extra, "Correlation_results_plots.tiff"), width=14, height=14, dpi=600)
```


```{r}
total_data.yearly <- total_data.yearly %>%
                                mutate(age=as.numeric(age)) %>%
                                mutate( location2 = location, location = ifelse(location2 == "Eastern Gotland Basin", "EGB", 
                                                                         ifelse(location2 == "Landsort Deep", "LD",
                                                                             ifelse(location2 == "Gulf of Finland", "GoF", NA)))) %>%
                              mutate(method =gsub("Dinosterl+dinostanols (µg/gTOC)", "Dinosterl & Dinostanols (µg/gTOC)",method))
#selected_colours_methods
plot <- ggplot(total_data.yearly, aes(y = values, x = age)) +
    geom_point(aes(color= method)) +
    geom_ribbon(data = my_predicted,
                mapping = aes(x = age, ymax = upper, ymin = lower, fill = dataset),
                alpha = 0.3, inherit.aes = FALSE) +
    geom_line(data = my_predicted,
              mapping = aes(y = fit, x = age, colour = dataset)) +
    labs(y = "Abundance proxy", x = "Year CE") +
    theme(legend.position = "bottom")+
    facet_grid(location ~factor(organism, c("Phytoplankton", "Dinoflagellata")), scales = "free")+
  labs(title="Abundance proxy dynamics of phytoplankton over time", 
       subtitle = "Dots reflect raw data",
       colour= "Methods")+ 
  coord_cartesian(ylim = c(0, 100))+ 
  scale_color_manual( values=hoiho2, 
                    labels = c("Biomonitoring", "C25:1 & C27:1 (µg/gTOC)", "Dinosterl & Dinostanols (µg/gTOC)", "sedaDNA Metabarcoding")) +
  scale_fill_manual(values=hoiho2)+
  figure_theme+guides(fill= "none") + 
  guides(colour = guide_legend(override.aes = list(size=3))) 


print(plot)
ggsave(plot, file=file.path(output_path, "Method_comparison__biomarker_metbarcoding__MUC__GAM_all.png"), height=8, width=14)
ggsave(plot, file=file.path(output_path, "Method_comparison__biomarker_metbarcoding__MUC__GAM_all.tiff"), height=8, width=14, dpi=600)
write.table(my_predicted, file = file.path(output_path, "Method_comparison__biomarker_metbarcoding__MUC__GAM_all.tsv"), sep = "\t", row.names = F)
write.table(total_data.yearly, 
            file = file.path(output_path, "Method_comparison__biomarker_metbarcoding__MUC__measurements_all.tsv"), sep = "\t", row.names = F)

```

# Combine stats

```{r}
my_predicted

correlation_stats <-  stats_gam_predicted
correlation_stats <- correlation_stats %>% mutate( round.est = round(estimate, 2),
                                                   round.stats = round(statistic,2),
                                                   round.p.value = round( p.value,3)) 
write.table(correlation_stats, file = file.path(output_path, "Correlation_results_as_table_of_gam_n_interval.tsv"), sep = "\t", row.names = F)

# transform data for final table
correlation_stats_overview <- correlation_stats %>% mutate(Method1= gsub("(Metabarcoding|Biomonitoring).*","\\1", Method1),
                             Method2= gsub("(Metabarcoding|Biomonitoring).*","\\1", Method2),
                             timespan =gsub("(Monitoring|MUC).*","\\1", timespan),
                             estimate.p.value= ifelse(p.value <= 0.001, "***",
                                               ifelse( p.value < 0.01, "**",
                                               ifelse(p.value <= 0.05, "*","")))) %>%
                      mutate(estimate.p.value = paste(round.est,estimate.p.value, sep=""  ))%>%
  select(location, organism, timespan, Method1, Method2, dataset, stats_method, estimate.p.value)%>%
  pivot_wider(names_from = location, values_from = estimate.p.value)

write.table(correlation_stats_overview, 
            file = file.path(output_path, "Correlation_results_as_table_of_gam_n_interval__overview.tsv"), sep = "\t", row.names = F)

unique(correlation_stats_overview$Method1)
correlation_stats_overview <- correlation_stats_overview %>% 
        filter(Method1 =="Metabarcoding" |Method1 =="Dinosterl+dinostanols (µg/gTOC)" |Method1 =="C25:1+C27:1 (µg/gTOC)" )

write.table(correlation_stats_overview, 
            file = file.path(output_path, "Correlation_results_as_table_of_gam_n_interval__overview_relevant.tsv"), sep = "\t", row.names = F)

```
